Sun, Jan 28, 2024, 12:53 PM +07:00

+++ week_2/week_2.py
@@ -1,5 +1,5 @@
-x = int(input("What's x? "))
-y = int(input("What's y? "))
+x = float(input("What's x? "))
+y = float(input("What's y? "))
 
 z = x + y
